+++ backend/src/main/java/ua/edu/ukma/LibraryManager/services/PrincipalService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface PrincipalService {
 
-    Optional<Principal> getPrincipalById(Long principalId);
+    Optional<Principal> getPrincipalById(Integer principalId);
 
-    boolean deletePrincipal(Long principalId);
+    boolean deletePrincipal(Integer principalId);
 }
